Favorite complete

+++ app/src/main/java/com/kaisho/carforrent/model/CarsModel.kt
@@ -4,10 +4,10 @@ class CarsModel {
     var name: String? = null
     var image: String? = null
     var description: String? = null
-    var price: Int? = null
+    var price: Double? = null
 
     constructor()
-    constructor(name: String?, image: String?, description: String?, price: Int?) {
+    constructor(name: String?, image: String?, description: String?, price: Double?) {
         this.name = name
         this.image = image
         this.description = description
